requirements.txt
streamlit
python-dotenv
pdfplumber
python-docx
langchain
langchain-community
langchain-text-splitters
langchain-groq
sentence-transformers
faiss-cpu
groq
